Gather data on each driver from given club
This gathers the relevant data from the given club id and season ids and
for the given race week and season quarter. It stores it as a dict where
key is cust IDs and the values are lists of subsessions where those
cust_ids had races.

From here we can use this data to find the top in each category in each
series.